DOM helpers in dedicated file
