+++ src/Sim/Terminal/parsers.ts
@@ -1,6 +1,6 @@
 import jsonData from "../../Data/data.json" assert { type: "json" };
-import { getTheoryFromIndex } from "../../Utils/helpers.js";
-import { inputData } from "../main.js";
+import { inputData } from "../main";
+import { getTheoryFromIndex } from "../../Utils/helpers";
 
 export function parseSimParams(params: string[]) {
   if (params.length < 3) throw "At least 3 parameters [strat, sigma, start rho] required.";
