+++ src/Sim/main.ts
@@ -1,25 +1,26 @@
 import jsonData from "../Data/data.json" assert { type: "json" };
-import { qs, sleep, getTheoryFromIndex, logToExp, convertTime, formatNumber, defaultResult } from "../Utils/helpers.js";
-import { parseData } from "./parsers.js";
-import { getStrats } from "./strats.js";
-import t1 from "../Theories/T1-T8/T1.js";
-import t2 from "../Theories/T1-T8/T2.js";
-import t3 from "../Theories/T1-T8/T3.js";
-import t4 from "../Theories/T1-T8/T4.js";
-import t5 from "../Theories/T1-T8/T5.js";
-import t6 from "../Theories/T1-T8/T6.js";
-import t7 from "../Theories/T1-T8/T7.js";
-import t8 from "../Theories/T1-T8/T8.js";
-import wsp from "../Theories/CTs/WSP.js";
-import sl from "../Theories/CTs/SL.js";
-import ef from "../Theories/CTs/EF.js";
-import csr2 from "../Theories/CTs/CSR2.js";
+import { sleep, getTheoryFromIndex, logToExp, convertTime, formatNumber, defaultResult } from "../Utils/helpers";
+import { qs } from "../Utils/DOMhelpers";
+import { parseData } from "./parsers";
+import { getStrats } from "./strats";
+import t1 from "../Theories/T1-T8/T1";
+import t2 from "../Theories/T1-T8/T2";
+import t3 from "../Theories/T1-T8/T3";
+import t4 from "../Theories/T1-T8/T4";
+import t5 from "../Theories/T1-T8/T5";
+import t6 from "../Theories/T1-T8/T6";
+import t7 from "../Theories/T1-T8/T7";
+import t8 from "../Theories/T1-T8/T8";
+import wsp from "../Theories/CTs/WSP";
+import sl from "../Theories/CTs/SL";
+import ef from "../Theories/CTs/EF";
+import csr2 from "../Theories/CTs/CSR2";
 import fi from "../Theories/CTs/FI";
-import fp from "../Theories/CTs/FP.js";
-import rz from "../Theories/CTs/RZ.js";
-import mf from "../Theories/CTs/MF.js";
-import bap from "../Theories/CTs/BaP.js";
-import bt from "../Theories/Unofficial-CTs/BT.js";
+import fp from "../Theories/CTs/FP";
+import rz from "../Theories/CTs/RZ";
+import mf from "../Theories/CTs/MF";
+import bap from "../Theories/CTs/BaP";
+import bt from "../Theories/Unofficial-CTs/BT";
 import tc from "../Theories/Unofficial-CTs/TC";
 
 const output = qs(".output");
