+++ src/Utils/helpers.ts
@@ -1,18 +1,5 @@
 import jsonData from "../Data/data.json" assert { type: "json" };
 
-/** Raises an exception */
-const raise = (err: string) => {
-  throw new Error(err);
-};
-/** Alias of document.querySelector */
-export const qs = <T extends HTMLElement>(name: string) => document.querySelector<T>(name) ?? raise(`HtmlElement ${name} not found.`);
-/** Alias of document.querySelectorAll */
-export const qsa = <T extends HTMLElement>(name: string) => document.querySelectorAll<T>(name);
-/** Alias of document.createElement */
-export const ce = <T extends HTMLElement>(type: string) => (document.createElement(type) as T) ?? raise(`HtmlElement ${type} could not be created.`);
-/** Adds an event listener to `element` */
-export const event = <T>(element: HTMLElement, eventType: string, callback: (e: T) => void) => element.addEventListener(eventType, (e) => callback(e as T));
-
 /** Returns the index of the first occurrence of `val` in `arr`, or -1 if not found */
 export function findIndex<T>(arr: T[], val: T): number {
   for (let i = 0; i < arr.length; i++) if (val === arr[i]) return i;
